weather_producer/requirements.txt
azure-eventhub
dotenv